+++ app/src/main/java/net/azurewebsites/streambeta/yandexstreamsandroid/domain/presenter/interfaces/DonateFragmentPresenter.java
@@ -9,15 +9,11 @@ import net.azurewebsites.streambeta.yandexstreamsandroid.domain.model.dto.Stream
 import net.azurewebsites.streambeta.yandexstreamsandroid.domain.presenter.instancestate.DonateFragmentPresenterInstanceState;
 import net.azurewebsites.streambeta.yandexstreamsandroid.domain.router.MainRouter;
 
-/**
- * Created by djavid on 30.07.17.
- */
-
 
 public interface DonateFragmentPresenter
         extends Presenter<DonateFragmentView, MainRouter, DonateFragmentPresenterInstanceState> {
 
-    void loadStreamSettings(int stream_id);
+    void loadStreamSettings(long stream_id);
     void requestPayment();
     void processPayment(RequestPayment requestPayment);
     void sendDonationWithId(DonationDto donationDto);
