+++ app/src/main/java/net/azurewebsites/streambeta/yandexstreamsandroid/domain/router/MainRouter.java
@@ -8,6 +8,6 @@ import net.azurewebsites.streambeta.yandexstreamsandroid.core.Router;
 
 public interface MainRouter extends Router {
     void goToScreen(ScreenTag screenTag);
-    void goToStreamPage(int streamId, String streamName, String streamerId);
+    void goToStreamPage(long streamId, String streamName, String streamerId);
     ScreenTag getCurrentScreen();
 }
